+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- All `DistributedGraph` config options (e.g. `grpc_options`, `num_threads`, ...) are exposed to `DistributedClient` and `BackendOptions`
+
 ## [0.1.58] - 2022-02-15
 
 ### Added